raytrace: in LBLATM the call when ITER=-1 (final or kb) uses IREAD=2 which is different than the call during the iteration (IREAD=1). Changed to also use IREAD=1 for ITER=-1.
frwdmdl: added a flag FDOCROSS. It is used to redo the raytrace (LBLATM) and cross sections (SETUP3) when ICOUNT=1. There was a problem with the SZA kb calculation when either temperature is retrieved or when kb temperature is requested. In both cases the SZA kb routine re-reads the temperature profile from the preference.prf. This leads to huge SZA errors when temperature is retrieved, or to a delta in temperature (spline for iread=1 vs weighted mean for iread=0) when kb temperature is requested. The delta Yn - YC for the kb calculation included not only the sza perturbation but also a delta in temperature. In the frwdmdl loop (icount=1) the YN is now calculated such that all subsequent perturbations (icount>1) are correct. If temperature is not retrieved but requested in kb, the YN is calculated with the spline interpolated temperature from the LBLATM (with iread=1) call. This is to ensure that temperature is not changed the the kb temperature components are perturbed. If kb sza is calculated without kb temperature, the old method is used (for backward compatibility...but maybe this is not necessary).