lineparam.f90: returning TAUMIN to the old values (o3 retrieval was a bit slow)
raytrace.f90: printing to sfit4.dtl is also depending on NOPRNT (when temperature retrieval is on, too many print statements in the sfit4.dtl appeared)

sfit4.f90: kbcalculate: I had a large difference in FITRMS when using apriori polynomial coefficients different from 1 and when extended phase and apod are retrieved. The cause was that initialize will set EPHSF as ESPHSF0, but EPHSF contains the retrieved phase ... and then the phase went off when multiplying with the PARM values.